corrections

+++ Main.py
@@ -125,7 +125,7 @@ def train_model(unet, optimizer, scheduler, dataloader, dataset_sizes, device, l
                                 _, preds = torch.max(outputs, 1)
                                 print(preds.shape)
                                 print(labels.shape)
-                                loss = F.binary_cross_entropy(outputs.type(torch.LongTensor), labels)
+                                loss = F.binary_cross_entropy(outputs, labels)
                                 # backward + optimize only if in training phase
                                 if phase == 'train':
                                     loss.backward()
